rag-qwen-faiss/requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
langchain
pymupdf
numpy
tqdm